program/localRAGV1.0.1/requirements.txt
langchain
langchain-community
langchain-ollama
langchain_huggingface
faiss-cpu
docling
sentence-transformers